adicionando confirmação para a remoção de tarefas

+++ funcionalidades.js
@@ -1,17 +1,23 @@
 var tarefas = [];
 
 function removerTarefa(posicao) {
-    tarefas.splice(posicao, 1);
-    localStorage.setItem('tarefas', JSON.stringify(tarefas))
 
-    document.querySelector('#tarefas').innerHTML = '';
-    tarefas.forEach((tarefa, posicao) => {
-        document.querySelector('#tarefas').innerHTML += `
-            <div class="tarefa" onClick="removerTarefa(${posicao})">
-                <span>${tarefa}</span>
-            </div>
-        `;
-    });
+    // Retorna True ou False dependendo da ação do usuário.
+    var confirmacao = confirm("Você deseja mesmo remover essa tarefa?")
+
+    if(confirmacao) {
+        tarefas.splice(posicao, 1);
+        localStorage.setItem('tarefas', JSON.stringify(tarefas))
+    
+        document.querySelector('#tarefas').innerHTML = '';
+        tarefas.forEach((tarefa, posicao) => {
+            document.querySelector('#tarefas').innerHTML += `
+                <div class="tarefa" onClick="removerTarefa(${posicao})">
+                    <span>${tarefa}</span>
+                </div>
+            `;
+        });
+    }
 }
 
 function recuperaTarefas() {
